Extract nonprofit state to separate file; add functionality to get GPS coords for the inputted zip code; get nonprofit data inside of setInputZip instead of requiring a separate call
